LibWeb: Split ShadowStyleValue out of StyleValue.{h,cpp}
